+++ src/synthv1_formant.h
@@ -75,8 +75,8 @@ public:
 		// update method
 		void update(float cutoff = 0.5f, float reso = 0.0f)
 		{
-			if (::fabs(m_cutoff - cutoff) > 0.001f ||
-				::fabs(m_reso   - reso)   > 0.001f) {
+			if (::fabsf(m_cutoff - cutoff) > 0.001f ||
+				::fabsf(m_reso   - reso)   > 0.001f) {
 				m_cutoff = cutoff;
 				m_reso = reso;
 				reset_coeffs();
@@ -225,8 +225,8 @@ protected:
 	// update method
 	void update(float cutoff, float reso)
 	{
-		if (::fabs(m_cutoff - cutoff) > 0.001f ||
-			::fabs(m_reso   - reso)   > 0.001f) {
+		if (::fabsf(m_cutoff - cutoff) > 0.001f ||
+			::fabsf(m_reso   - reso)   > 0.001f) {
 			m_cutoff = cutoff;
 			m_reso = reso;
 			reset_coeffs();
